+++ integrations/krita/pykrita/cel_shaded_generator/__init__.py
@@ -2,6 +2,7 @@
 
 from krita import DockWidgetFactory, DockWidgetFactoryBase, Krita
 
+from .chapter_docker import ChapterQueueDocker
 from .color_docker import CharacterColorsDocker
 from .docker import LearningDocker
 from .segmentation_docker import SegmentationDocker
@@ -25,5 +26,12 @@ Krita.instance().addDockWidgetFactory(
         "cel_shaded_generator_segmentation_docker",
         DockWidgetFactoryBase.DockRight,
         SegmentationDocker,
+    )
+)
+Krita.instance().addDockWidgetFactory(
+    DockWidgetFactory(
+        "cel_shaded_generator_chapter_queue_docker",
+        DockWidgetFactoryBase.DockRight,
+        ChapterQueueDocker,
     )
 )